# openapi.yml
openapi: 3.0.3
info:
  title: Auction Hub API
  description: |
    Complete API documentation for the Auction Hub backend system.

    ## Features
    - User authentication and role-based access control
    - Two-tier auction registration approval system
    - Real-time bidding via WebSocket
    - Stripe payment integration for deposits and winner payments
    - Auction policy management and fee calculations
    - Comprehensive audit trails

    ## Authentication
    Most endpoints require JWT Bearer token authentication.
    Include the token in the Authorization header: `Bearer <token>`

    ## User Roles
    - `bidder` - Default role, can bid on auctions
    - `auctioneer` - Can manage auctions and registrations
    - `admin` - Full access except super admin functions
    - `super_admin` - Complete system access

  version: 3.0.0
  contact:
    name: Auction Hub Support
    email: support@auctionhub.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.auctionhub.com/api
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
...
paths:
...
  /system-variables:
    get:
      tags:
        - System Variables
      summary: Get all system variables
      description: View all system configuration variables
      operationId: getAllSystemVariables
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
      responses:
        '200':
          description: System variables
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - System Variables
      summary: Create system variable
      description: Create a new system configuration variable
      operationId: createSystemVariable
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - category
                - key
                - value
                - dataType
              properties:
                category:
                  type: string
                key:
                  type: string
                value:
                  type: string
                dataType:
                  type: string
                  enum: [number, boolean, string, json]
                description:
                  type: string
      responses:
        '201':
          description: Variable created
        '409':
          description: Variable already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /system-variables/{category}/{key}:
    get:
      tags:
        - System Variables
      summary: Get single variable
      description: Get a specific system variable value
      operationId: getSystemVariable
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Variable value
          content:
            application/json:
              schema:
                type: object
                properties:
                  category:
                    type: string
                  key:
                    type: string
                  value:
                    type: string

    patch:
      tags:
        - System Variables
      summary: Update variable
      description: Update system variable value
      operationId: updateSystemVariable
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: path
          required: true
          schema:
            type: string
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - value
              properties:
                value:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Variable updated
...
  /system-variables/cache/clear:
    post:
      tags:
        - System Variables
      summary: Clear cache
      description: Clear system variables cache
      operationId: clearCache
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Clear cache for specific category only
      responses:
        '200':
          description: Cache cleared
  /system-variables/cache/stats:
    get:
      tags:
        - System Variables
      summary: Get cache statistics
      description: View cache performance metrics
      operationId: getCacheStats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Cache statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalEntries:
                    type: integer
                  validEntries:
                    type: integer
                  expiredEntries:
                    type: integer
                  cacheTTL:
                    type: integer

  # ==================== AUCTION COSTS ====================
...